+++ lib/openapi_first/test/methods.rb
@@ -7,6 +7,16 @@ module OpenapiFirst
   module Test
     # Methods to use in integration tests
     module Methods
+      def self.[](*)
+        mod = Module.new do
+          def self.included(base)
+            OpenapiFirst::Test::Methods.included(base)
+          end
+        end
+        mod.define_method(:app) { OpenapiFirst::Test.app(*) }
+        mod
+      end
+
       def self.included(base)
         if Test.minitest?(base)
           base.include(OpenapiFirst::Test::MinitestHelpers)